refactor(remote): reduce vmux_remote to the transport contract
The relay links this crate to move bytes between a phone and a desktop, and
today that means it also compiles in Message, RoomEvent, AssistantBlock and
every session view — so a relay could decode a payload, and the only thing
stopping it is that nobody wrote the call.

Move the domain half to vmux_wire::room and drop the vmux_wire dependency
entirely. What is left is what the relay legitimately needs: DeviceId, the
desktop command envelope, and the QUIC setup types. The property is now
structural rather than a matter of restraint.

Adds the connection contract the transport work needs: ALPN, a fixed-layout
ClientHello carrying protocol_version, capabilities and resume_from, StreamKind
and close codes. The hello is JSON behind a magic-and-length prelude rather
than rkyv, because it is read before the peers have agreed a version and rkyv
would misread a reordered variant instead of noticing the mismatch. Unknown
capabilities degrade to Unknown so a newer client is understood, not refused.

Locks the rkyv discriminant order behind a v1 byte fixture, since that is the
one thing a phone shipping on its own schedule can silently break. Verified by
mutation: moving a variant turns it red, and it still compiles, so the failure
is the guard rather than the build.
