+++ crates/vmux_remote/Cargo.toml
@@ -2,14 +2,14 @@
 name = "vmux_remote"
 version.workspace = true
 edition.workspace = true
-description = "Shared protocol types for Vmux Remote clients"
+description = "Transport contract between Vmux clients, the relay and a desktop"
 publish = false
 
+# Deliberately depends on no vmux crate. The relay links this to route bytes; keeping the domain
+# types out means it cannot decode a payload even by accident.
 [dependencies]
 serde = { workspace = true }
 serde_json = { workspace = true }
-unicode-segmentation = "1"
-vmux_wire = { path = "../vmux_wire" }
 
 [dev-dependencies]
 serde_json = { workspace = true }
